+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+#### 0.1.1 - 30.08.2016
+* Fix NuGet package
 #### 0.1.0 - 27.08.2016
 * Add basic API for deploying Azure WebApps